backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asn1crypto==1.5.1
attrs==25.4.0
base58==2.1.1
bcl==2.3.1
bcrypt==5.0.0
bip_utils==2.10.0
bitarray==3.8.0
blinker==1.9.0
cachetools==6.2.4
cbor2==5.7.1
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
coinbase-agentkit
coinbase-agentkit-langchain
coincurve==20.0.0
construct==2.10.70
construct-typing==0.7.0
crcmod==1.7
cryptography==44.0.3
cytoolz==1.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
ed25519-blake2b==1.4.1
egcd==2.0.2
email-validator==2.3.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.124.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
Flask==3.1.2
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonalias==0.1.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lagrange==3.0.1
langchain==0.3.27
langchain-core==0.3.80
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
langgraph==0.2.76
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.1
langgraph-sdk==0.1.74
langsmith==0.4.59
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
motor==3.7.1
multidict==6.7.0
nest-asyncio==1.6.0
nilql==0.0.0a13
numpy==2.4.0
openai==1.109.1
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pailliers==0.3.0
pandas==2.3.3
pandas-ta-classic==0.3.59
paramiko==3.5.1
parsimonious==0.10.0
propcache==0.4.1
py-sr25519-bindings==0.2.3
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.5
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.3
rabinmiller==0.2.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rlp==4.1.0
rpds-py==0.30.0
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
solana==0.36.10
solders==0.27.1
SQLAlchemy==2.0.45
starlette==0.50.0
tavily-python==0.7.17
tenacity==9.1.2
tiktoken==0.12.0
toolz==1.1.0
tqdm==4.67.1
typer==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
web3==7.10.0
websockets==15.0.1
Werkzeug==3.1.4
x402==1.0.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
